improvements #17

+++ StepMaster/StepMaster/Models/MultidayStepsAchievement.cs
@@ -10,6 +10,9 @@ namespace StepMaster.Models
         public int NumberOfDays { get; set; }
         public override bool Check(List<int> steps)
         {
+            if (steps.Count < 7 - NumberOfDays)
+                return false;
+
             steps.RemoveRange(0, 7 - NumberOfDays);
 
             foreach (int s in steps)
